useful modules/requirements.txt
async-generator==1.10
attrs==21.4.0
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
futures==3.0.5
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
kaitaistruct==0.9
loguru==0.6.0
oauthlib==3.2.0
outcome==1.1.0
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.1.2
selenium-wire==4.6.2
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.6.0
urllib3==1.26.8
webdriver-manager==3.5.3
wsproto==1.1.0
zstandard==0.17.0
